src/test/java/ru/example/home/crud/controller/MainControllerTest.java
package ru.example.home.crud.controller;

import static org.assertj.core.api.Assertions.assertThat;

import org.junit.After;
import org.junit.Before;
import org.junit.jupiter.api.Test;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;

import static org.junit.Assert.*;
@SpringBootTest
public class MainControllerTest {

    @Autowired
    private MainController mainController;
    @Before
    public void setUp() throws Exception {
    }

    @After
    public void tearDown() throws Exception {
    }

    @Test
    public void getAllPersons() {
        assertThat(mainController).isNotNull();
    }
}